+++ src/commands/balance.rs
