+++ src/commands/prices.rs
