+++ src/commands/register.rs
