+++ src/mod.rs
@@ -2,14 +2,14 @@ extern crate pest;
 #[macro_use]
 extern crate pest_derive;
 
+mod app;
 pub mod commands;
 mod error;
 pub mod filter;
 mod list;
-mod main;
 pub mod models;
 pub mod parser;
 
+pub use app::{run_app, CommonOpts};
 pub(crate) use error::{Error, LedgerError, ParserError};
 pub use list::List;
-pub use main::{run_app, CommonOpts};
